Add a primitive low-level HTTP server.
This low-level HTTP server enables us to not have to manually parse
HTTP packets ourselves without adding any middleware, routing or SSL
layer overhead. It makes use of asyncio and aiohttp (itself built on
asyncio) to provide a first-class experience in this regard.

I have additionally added a Makefile to simplify local development.

There is an argument to be made that it seems foolish to add an HTTP
server behind Envoy when Envoy can support arbitrary TCP-based
protocols. The rationale for not going that way is that it ensures
simplicity in the POC phase.