Introduce saga coordination logic for the first time.
This commit:

- Drops the state machine model introduced in
https://github.com/Brown-CS2952-cwu-amahajan/qbox/pull/2. Instead, we
implement a raw execution engine that is closely coupled to
configuration specifics.

- Revamps the configuration model. Instead of specifying states for the
state machine, we introduce conventions to match requests, a template
for generating requests, and guarantees on header inheritance by
compensating transactions.

- Introduces a `SagaCoordinator` class that fully respects timeouts, and
can issue compensating transactions when some response fails.

- Introduces unit tests and schema validation for configuration as well
as the `SagaCoordinator` class, including network requests mocking.

This is a fairly large commit, but our work isn't done yet:

- We still need to figure out traffic redirection logic between Envoy, application
and the qbox in the next two weeks.

- The tests are incomplete, and only test basic saga logic. It doesn't
test any of the guarantees we will actually need, such as ensuring
headers propagate correctly from parent transaction to child
compensating transaction.

- There is no suppot for asynchronous saga requests. Everything is
synchronous or nothing. We should revisit this at a later date - for
now, I am fine just keeping everything synchronous for simplicity.

- We still need to build out the `handler` logic in `server.py` to
interface with the `SagaCoordinator`. One thing I want to do is either
create a new process to run `SagaCoordinator` in or add a coordinator to
our event loop, so that the main loop isn't blocked.